refactor(charts): replace SimpleCharts shared barrel with direct imports
Remove index.ts re-export layer. Each component now imports directly
from the source file it needs.
