+++ app/src/components/Charts/SimpleCharts/HourlyStreams/HourlyStreams.tsx
@@ -1,7 +1,10 @@
 import type { FC } from 'react'
 import { useState } from 'react'
 import type { HourlyStreamsQueryResult } from './query'
-import { ChartCard, ChartCardEmpty, ChartHero, ChartTooltip } from '../shared'
+import { ChartCard } from '../shared/ChartCard'
+import { ChartCardEmpty } from '../shared/ChartCardEmpty'
+import { ChartHero } from '../shared/ChartHero'
+import { ChartTooltip } from '../shared/ChartTooltip'
 import { formatDuration } from '../../../../utils/formatDuration'
 import { clockEmoji } from './clockEmojis'
