+++ app/src/components/Charts/SimpleCharts/ListeningRhythm/ListeningRhythm.tsx
@@ -1,11 +1,9 @@
 import type { FC } from 'react'
 import type { ListeningRhythmResult } from './query'
-import {
-    ChartCard,
-    ChartCardEmpty,
-    ChartHero,
-    LabeledProgressBar,
-} from '../shared'
+import { ChartCard } from '../shared/ChartCard'
+import { ChartCardEmpty } from '../shared/ChartCardEmpty'
+import { ChartHero } from '../shared/ChartHero'
+import { LabeledProgressBar } from '../shared/LabeledProgressBar'
 
 type Props = {
     data: ListeningRhythmResult | undefined
