+++ app/src/components/Charts/SimpleCharts/Regularity/Regularity.tsx
@@ -1,7 +1,9 @@
 import type { FC } from 'react'
 import type { RegularityResult } from './query'
 import { classifyRegularity } from './classifyRegularity'
-import { ChartCard, ChartCardEmpty, ChartHero } from '../shared'
+import { ChartCard } from '../shared/ChartCard'
+import { ChartCardEmpty } from '../shared/ChartCardEmpty'
+import { ChartHero } from '../shared/ChartHero'
 
 type Props = {
     data: RegularityResult | undefined
