+++ app/src/components/Charts/SimpleCharts/SessionAnalysis/SessionAnalysis.tsx
@@ -1,6 +1,9 @@
 import type { FC } from 'react'
 import type { SessionAnalysisResult } from './query'
-import { ChartCard, ChartCardEmpty, ChartHero, InsightCard } from '../shared'
+import { ChartCard } from '../shared/ChartCard'
+import { ChartCardEmpty } from '../shared/ChartCardEmpty'
+import { ChartHero } from '../shared/ChartHero'
+import { InsightCard } from '../shared/InsightCard'
 import { formatDuration } from '../../../../utils/formatDuration'
 
 type Props = {
