--- app/src/components/Charts/SimpleCharts/shared/index.ts
@@ -1,9 +1,0 @@
-export { ChartCard } from './ChartCard'
-export { ChartHero } from './ChartHero'
-export { rankEmojis } from './rankEmojis'
-export { RankedList, type RankedItem } from './RankedList'
-export { ProgressBar } from './ProgressBar'
-export { LabeledProgressBar } from './LabeledProgressBar'
-export { InsightCard } from './InsightCard'
-export { ChartTooltip } from './ChartTooltip'
-export { ChartCardEmpty } from './ChartCardEmpty'
